Got ANF with multi-language boundaries working
New multi-language pattern that actually uses boundaries.

The strategy is to define a "translation context", under translation reductions
such as the A-reduction trigger.
This defines the context for the context closure of the translation reduction.
The translation context extends the compatible closure of the target language,
but requires an intervening TS boundary.
I.e., the context is (in-hole C (TS C)).
This means the term in the hole is a "source" term, since it's on the S side of
the boundary.

The translation reduction should insert new boundaries ST and TS boundaries to
demark source/target transitions, assuming they are already under a single TS
boundary.

The reduction relation is also extended with the usual two administrative
redexes: boundaries can be collapsed under any context.