Fix translation context, via Lily
Looks like the translation context can't pick an arbitrary subterm, but must
translate the entire term under the boundary?
Seems too restrictive, but now we get nice confluence diagrams in the testing
file added by this patch.
Maybe need to consider how the translation context relates to the evaluation
context that the A reduction manipulate.